Override grapheme cluster width
Add functionality to override the grapheme cluster width. This is
useful when dealing with terminals that do not confirm to the Unicode
specification and display a grapheme cluster with the incorrect width.

By having uniseg report back the same response as the terminal, this
allows for the terminal to match cursor positioning with the
application. While the results are incorrect, this still provides a
better user experience.

Signed-off-by: Michael Lorant <michael.lorant@nine.com.au>